+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 ### Added
 
 - [go-kit](https://github.com/go-kit/kit/tree/master/log) logger constructor
+- [emperror](https://github.com/goph/emperror) error handler constructor